pyproject.toml
[project]
name = "sdpo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "huggingface-hub>=1.3.7",
    "torch>=2.10.0",
    "transformers>=5.0.0",
    "openai>=1.0.0",
    "langdetect>=1.0.9",
    "accelerate>=1.12.0",
    "wandb>=0.24.1",
    "python-dotenv>=1.2.1",
    "immutabledict>=4.2.2",
    "nltk>=3.9.2",
]
